Adding support for forwarding regfile - abandoned
This commit adds support for a forwarding regfile. The support is not
fully working and still has issues with epochs changing and meaning that
instructions get the wrong forwarded values.

This has been abandoned because it does not provide an improvement to
the critical path and is quite large in terms of area.