Delayed memory reqs, fix BE & sign extension bugs
This commit adds a macro that allows delaying the memory requests for a
cycle in order to allow a more accurate view of the CPU's performance
memory system logic being counted.
This commit also fixes a bug with both the byteenable signal of the
Avalon memory and correctly sign-extends memory loads.