Change bypassing logic to improve fmax
This commit moves the bypassing logic from the start of Stage1 to the
end of StageD, in Stage1's enq method. This means the ALU operations are
not dependent on the bypassing.
This does require an extra cycle for memory operations in the current
setup in which memory operations only bypass once they are in stage3.
However, it is possible to change this so that they can bypass while
still in Stage2.

The changes in this commit also have the side-effect of preventing the
register file from being inferred as an altsyncram in Quartus.