Decompress in StageF, add bypass deduplication
This moves instruction decompression into StageF in order to allow the
register file to be inferred as an altsyncram in Quartus.
This commit also adds a `define which allows the bypass logic to be
deduplicated. Deduplicating means that the design will be slightly
smaller but will have a lower fmax.
Not deduplicating leads to ~7% higher performance and ~5% higher area
usage on Cyclone V.