Squashed past few commits to view total changes
This commit is just a squashed version of the past few commits in order
to see the full changes that have occurred.

Add Avalon memory, inline ALU and disable BP

This commit adds a stub Avalon memory interface (from RISC-V Flute
implementation), inlines functions in the ALU, and disables branch
prediction. This commit also adds some synthesize boundaries, and
adds a new signal from Stage1 to Stage2 to remove getBase(authority) and
getTop(authority) in Stage 2.

(WIP!!) Changes to improve size & fmax

NOTE: THIS IS A WORK-IN-PROGRESS COMMIT. THIS WILL BE AMENDED AND
FORCE-PUSHED.

There are several changes in this commit. The major ones are:
- A significant amount of the decoding logic has been moved from the ALU
  to the decode stage.
- New register bypassing has been added to reduce the length of the
  critical path in the ALU.
- Multiplication has an extra cycle of delay (to improve fmax)
- The implementation of PCC_T has been changed (the only major change
  that is visible to the outside are that toCapPipeNoOffset is cheaper
  than toCapPipe but does not set the offset (instead, the address
  field stores the base of PCC)

TODO include more changes here

More changes to ALU, Stage1, CPU; read description

This commit has the following changes:
- Removes all Stage1 traps. This means the rule that deals with stage1
  traps has also been removed from CPU.bsv
- Makes trap data be passed from stage1 to stage2 (rather than just
  being output directly to CPU.bsv) so that stage2 has trap information.
- Changes toCapPipe to toCapPipeNoOffset in CPU.bsv to avoid an
  expensive operation.
- slightly changes the implementation of PCC_T to make it easier to read
  and change. It also changes the implementation of toCapPipe to be
  cheaper but unsafe. TODO: check whether it is still OK to make it
  unsafe
- Adds a new DISABLE_BRANCH_PRED define which removes some signals that
  were used only for branch prediction
  TODO wrap the actual branch prediction in stageF in this
- Adds some new signals from the decoder to stage1 so that AUIPCC and
  JALR function correctly
- Changes the Stage1 & ALU so that it mostly uses the PCC's address
  rather than the offset (which makes it faster & cheaper).
- Cleans up a bunch of comments in the ALU that were being used for
  reference in case the previous implementation was required.
- Muxes inputs to fromCapPipe in order to reduce area usage in the ALU.

Initial changes to add BRAM for CSRs and SCRs

This adds a BRAM in CSR_RegFile_MSU so that CSRs and SCRs can eventually be
stored in there, saving area on FPGA.
The BRAM is correctly inferred in Quartus.
Currently, only the uncommented registers in ISA_Decls_CSR use the BRAM,
and these have not yet been removed from the CSR_RegFile_MSU so for now
they are duplicated and area usage is higher.

Move rest of SCRs into BRAM & other optimisations

This commit moves the SCRs that had not yet been moved into the BRAM.
It also does some more optimisations, including:
- Get rid of an unnecessary access_permitted port
- Move the actual CSR reading to the start of the cycle after the
  request is made (to improve fmax)
- Get rid of rg_csr_scr_address in CPU.bsv since we can just work it out
  from the trap information

Move some CSRs to CSR/SCR BRAM, CSR cleanup

This commit moves the tselect, tdata1, tdata2 and tdata3 CSRs to the CSR
BRAM.
This commit also does some cleanup in the CSR reset logic to remove one
unnecessary cycle from the reset, and adds some comments