Initial changes to add BRAM for CSRs and SCRs
This adds a BRAM in CSR_RegFile_MSU so that CSRs and SCRs can eventually be
stored in there, saving area on FPGA.
The BRAM is correctly inferred in Quartus.
Currently, only the uncommented registers in ISA_Decls_CSR use the BRAM,
and these have not yet been removed from the CSR_RegFile_MSU so for now
they are duplicated and area usage is higher.