Changes to add BRAM for CSRs and SCRs
This adds a BRAM in CSR_RegFile_MSU so that CSRs and SCRs can eventually be
stored in there, saving area on FPGA.
The BRAM is correctly inferred in Quartus.
Currently, only the uncommented registers in ISA_Decls_CSR use the BRAM,
and these have not yet been removed from the CSR_RegFile_MSU so for now
they are duplicated and area usage is higher.

Move rest of SCRs into BRAM & other optimisations

This commit moves the SCRs that had not yet been moved into the BRAM.
It also does some more optimisations, including:
- Get rid of an unnecessary access_permitted port
- Move the actual CSR reading to the start of the cycle after the
  request is made (to improve fmax)
- Get rid of rg_csr_scr_address in CPU.bsv since we can just work it out
  from the trap information
- Move tselect, tdata1, tdata2, tdata3 to the SCR BRAM
- Clean up BRAM reset logic to remove an unnecessary cycle from the
  reset

Lastly this commit adds a couple of comments to the CSR Regfile